remove some warnings

+++ HLTriggerOffline/Scouting/plugins/PatElectronTagProbeAnalyzer.cc
@@ -53,7 +53,7 @@ void PatElectronTagProbeAnalyzer::dqmAnalyze(edm::Event const& iEvent,
   iEvent.getByToken(triggerObjects_, triggerObjects);
 
   std::vector<std::string> filterToMatch = {"hltDoubleEG11CaloIdLHEFilter", "hltEG30EBTightIDTightIsoTrackIsoFilter"};
-  int numberOfFilters = filterToMatch.size();
+  size_t numberOfFilters = filterToMatch.size();
   trigger::TriggerObjectCollection* legObjects = new trigger::TriggerObjectCollection[numberOfFilters];
   for (size_t iteFilter = 0; iteFilter < filterToMatch.size(); iteFilter++) {
     std::string filterTag = filterToMatch.at(iteFilter);
